apps/relay/python/requirements.txt
pyirsdk
python-socketio
aiohttp
